% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calender_data.R
\name{calender_data}
\alias{calender_data}
\title{Make calendar data}
\usage{
calender_data(
  month = Sys.Date(),
  headline_factor = 2,
  start_monday = TRUE,
  locale = NULL
)
}
\arguments{
\item{month}{either a date or name of a month in the current locale.}

\item{headline_factor}{the factor by which the name of the month is larger
then the remaining text (2 means twice the size).}

\item{start_monday}{If TRUE the week starts on Monday, otherwise starts on
Sunday.}

\item{locale}{provide locale if you want to use non English names for months
and days.}
}
\value{
data.frame. Calendar data that works well with ggplot2.
}
\description{
Make the calendar data for plotting. Can be used alone if you want a more
specialised plotting experience.
}
