% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_calendar_background.R
\name{make_calendar_background}
\alias{make_calendar_background}
\title{Make Calendar Background}
\usage{
make_calendar_background(
  img,
  filename,
  month = Sys.Date(),
  resolution = "auto",
  position = "center",
  scale = TRUE,
  colour = "white",
  fill = "grey",
  family = "",
  text_size = 9,
  headline_factor = 2,
  start_monday = TRUE,
  locale = NULL
)
}
\arguments{
\item{img}{character. Location of a background image.}

\item{filename}{character. File name of the output.}

\item{month}{either a date or name of a month in the current locale.}

\item{resolution}{"auto" to use the resolution of the input image or vector
of length two with width and height of the output image.}

\item{position}{position of calendar in the image. Possible values are:
"center", "bottom", "top", "right", "left", "bottomright", "bottomleft",
"topright", "topleft" and a numeric vector of length 4 .}

\item{scale}{if resolution is provided, should the image be scaled down/up
before cropping to the provided resolution?}

\item{colour}{colour of the font.}

\item{fill}{a colour to fill the background.}

\item{family}{font.}

\item{text_size}{size of the text in the image.}

\item{headline_factor}{the factor by which the name of the month is larger
then the remaining text (2 means twice the size).}

\item{start_monday}{If TRUE the week starts on Monday, otherwise starts on
Sunday.}

\item{locale}{provide locale if you want to use non English names for months
and days.}
}
\value{
saves image in file location
}
\description{
Make Calendar Background
}
\examples{
make_calendar_background(
  system.file("extdata", "mull.jpg", package = "wallpapr")
)

# put in upper right corner
make_calendar_background(
  system.file("extdata", "zima.png", package = "wallpapr"),
  position = c(2, 0, 2, 0)
)
}
