% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_wallpapr.R
\name{save_wallpapr}
\alias{save_wallpapr}
\title{Save wallpapr as a file}
\usage{
save_wallpapr(plot, filename)
}
\arguments{
\item{plot}{plot created with make_wallpapr.}

\item{filename}{character. File name of the output.}
}
\description{
Export the plot created with make_wallpapr to an image file.
}
