% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_calendar_background.R
\name{calender_data}
\alias{calender_data}
\title{Make calendar data}
\usage{
calender_data(month, headline_factor, start_monday, locale)
}
\arguments{
\item{month}{either a date or name of a month in the current locale.}

\item{headline_factor}{the factor by which the name of the month is larger
then the remaining text (2 means twice the size).}

\item{start_monday}{If TRUE the week starts on Monday, otherwise starts on
Sunday.}

\item{locale}{provide locale if you want to use non English names for months
and days.}
}
\description{
Make the calendar data for plotting.
}
