% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_calendar_background.R
\name{make_calendar_background}
\alias{make_calendar_background}
\title{Make Calendar Background}
\usage{
make_calendar_background(img, filename = paste0(format(Sys.Date(), format
  = "\%B"), img_format(img)), month = Sys.Date(), resolution = "auto",
  colour = "white", fill = "grey", family = "", text_size = 9,
  headline_factor = 2, start_monday = TRUE, locale = NULL)
}
\arguments{
\item{img}{character. Location of a background image.}

\item{filename}{character. File name of the output.}

\item{month}{either a date or name of a month in the current locale.}

\item{resolution}{"auto" to use the resolution of the input image or vector
of length two with width and height of the output image.}

\item{colour}{colour of the font.}

\item{fill}{a colour to fill the background.}

\item{family}{font.}

\item{text_size}{size of the text in the image.}

\item{headline_factor}{the factor by which the name of the month is larger
then the remaining text (2 means twice the size).}

\item{start_monday}{If TRUE the week starts on Monday, otherwise starts on
Sunday.}

\item{locale}{provide locale if you want to use non English names for months
and days.}
}
\value{
saves image in file location
}
\description{
Make Calendar Background
}
\examples{
make_calendar_background(
  system.file("extdata", "mull.jpg", package = "wallpapr")
)
}
